+++ api/rust/include/LIEF/rust/PDB/types/Method.hpp
@@ -14,6 +14,7 @@
  */
 #pragma once
 #include "LIEF/rust/Mirror.hpp"
+#include "LIEF/rust/helpers.hpp"
 #include "LIEF/PDB/types/Method.hpp"
 
 class PDB_types_Method : private Mirror<LIEF::pdb::types::Method> {
@@ -22,4 +23,8 @@ class PDB_types_Method : private Mirror<LIEF::pdb::types::Method> {
   using lief_t = LIEF::pdb::types::Method;
 
   auto name() const { return get().name(); }
+
+  auto get_type() const { return to_int(get().type()); }
+
+  auto access() const { return to_int(get().access()); }
 };