+++ doc/sphinx/changelog.rst
@@ -35,6 +35,7 @@
   * Improve the support and the API for ``LF_PROCEDURE``: |lief-pdb-types-Function|
   * Improve the support and the API for ``LF_ARRAY``: |lief-pdb-types-Array|
   * Improve the support and the API for *simple* types: |lief-pdb-types-Simple|
+  * Improve the support and the API for ``LF_ONEMETHOD``: |lief-pdb-types-Method|
 
 :COFF: